bump versions and dependency versions
There was also a very weird bug where dylib symbols broke after their library get stored in a static `LazyCell<RwLock<HashMap<...>>>`, but they don't break if the library is retained via `std::mem::forget`. I thought there wouldn't be a difference, maybe there isn't supposed to be, but regardless `forget`ting is actually a better solution anyways because that was the entire purpose of the static storage.