Attempt to fix https://github.com/Jakobeha/type-sitter/issues/3
The fix is different than https://github.com/Jakobeha/type-sitter/pull/4. Instead of changing the punctuation names, if there would be multiple variants or methods with the same name in the same type, it adds underscores to the later one(s).

Although I only tested it a bit and didn't add any tests to the crate.