big refactor and API change
Anyone actually using this let me know if it breaks things. A lot of methods are renamed, but everything exposed before should still be exposed, and the new names make more sense.

In particular, a crate should now be able to just depend on `type-sitter-lib` and not `tree-sitter`, because everything is either wrapped or re-exported.