fix -Znext-solver issue and 2024 warning
The -Znext-solver issue is "expected generic lifetime parameter, found `'_`" with no source. It occurs as of Rust 1.86. It may be related to https://github.com/rust-lang/rust/issues/135152.

The -Znext-solver fix also makes documentation more deterministic.

The 2024 warning is that now we need `unsafe` blocks in `unsafe extern "C"`. functions.